Time: 2 ms (56%), Space: 46.2 MB (60.02%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.java
@@ -22,6 +22,6 @@ class Solution {
         }
         if((m+n)%2==1)
             return nums.get((m+n)/2);
-        return (float)(nums.get((m+n)/2)+nums.get((m+n)/2-1))/2;
+        return (nums.get((m+n)/2)+nums.get((m+n)/2-1))/2.0;
     }
 }
